seperering af note endpoints til notes.py med /notes prefix
