+++ hello_app/templates/createnote.html
@@ -19,7 +19,7 @@ Create Note
 <div class="main-body">
 <h1>Opret Note</h1>
 
-<form action="/create/" method="post">
+<form action="/notes/create/" method="post">
     <input type="text" name="title" placeholder="Name" /><br>
     <textarea name="note" placeholder="Note"></textarea><br>
     <!-- <input type="text" name="author" placeholder="Author" /><br> -->
